frontend: fixes unrelated failing test
This fixes random unrelated failing test (see error below).

It looks like importing event-legacy.ts in api/backends.ts causes
the issue (wrong initialization). balance.test.tsx includes rates.tsx
which imports api/backends.tsx

Having event-legacy.ts (or event.ts) imported in api/backend probably
break the mock in balance.test.tsx somehow.

- https://github.com/digitalbitbox/bitbox-wallet-app/blob/a73650bb05b091a7ff502a8766a6abf88727dcda/frontends/web/src/utils/event.ts#L38
- https://github.com/digitalbitbox/bitbox-wallet-app/blob/a73650bb05b091a7ff502a8766a6abf88727dcda/frontends/web/src/utils/event-legacy.ts#L37

There are a few other application relevant objects that probably all
should go into contexts.

Error message:

⎯⎯⎯⎯⎯⎯ Failed Suites 1 ⎯⎯⎯⎯⎯⎯⎯

 FAIL  src/components/balance/balance.test.tsx [ src/components/balance/balance.test.tsx ]
Error: [vitest] No "isTLS" export is defined on the "../../utils/request" mock. Did you forget to return it from "vi.mock"?
If you need to partially mock a module, you can use "vi.importActual" inside:

vi.mock("../../utils/request", async () => {
  const actual = await vi.importActual("../../utils/request")
  return {
    ...actual,
    // your mocked methods
  },
})

 ❯ Module.webSubscribePushNotifications src/utils/transport-websocket.ts:28:29
     26|   currentListeners.push(msgCallback);
     27|   if (!socket) {
     28|     socket = new WebSocket((isTLS() ? 'wss://' : 'ws://') + 'localhost…
       |                             ^
     29|
     30|     socket.onopen = function() {
 ❯ Module.apiWebsocket src/utils/websocket.ts:33:10
 ❯ src/utils/event-legacy.ts:77:1
 ❯ src/api/backend.ts:2:31
