+++ frontends/web/src/api/transactions.ts
@@ -1,0 +1,32 @@
+/**
+ * Copyright 2023 Shift Crypto AG
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import { subscribe as subscribeLegacy } from '../utils/event-legacy';
+
+export const syncNewTxs = (
+  cb: (
+    meta: {
+      count: number,
+      accountName: string,
+    }
+  ) => void,
+) => {
+  return subscribeLegacy('newTxs', event => {
+    if (event.type === 'backend') {
+      cb(event.meta);
+    }
+  });
+};
